// src/client/types.gen.ts
export type LoginInBody = {
    /**
     * A URL to the JSON Schema for this object.
     */
    readonly $schema?: string;
    handle_or_email: string;
    password: string;
};

export type SelfUser = {
    /**
     * A URL to the JSON Schema for this object.
     */
    readonly $schema?: string;
    admin: boolean;
    created_at: string;
    display_name: string;
    email: string;
    handle: string;
    id: string;
    verified: boolean;
};

export type SessionDto = {
    created_at: string;
    current: boolean;
    id: string;
    ip: string;
    last_seen: string;
    user_agent: string;
};

export type SignupInBody = {
    /**
     * A URL to the JSON Schema for this object.
     */
    readonly $schema?: string;
    /**
     * optional; defaults to the handle
     */
    display_name?: string;
    email: string;
    /**
     * lowercase letters, digits and underscore
     */
    handle: string;
    /**
     * 72 is bcrypt's limit
     */
    password: string;
};